Fix 697123: Handle nesting of block-boxes deep inside inline-boxes.
When we climb up the box tree to create a new block box for block-level
elements that are deep inside inline-level elements, we should propagate
this to the caller so it can also climb up. Otherwise other child
elements in the inline-level element will be inserted before the
block box that was created for the content that was shunted up the tree.