tiff: Set k>0 for CCITT fax group 3 1D/2D compression.
Previously, if a TIFF file was encoded using CCITT
fax group 3 compression and mixing 1D and 2D codes,
the fax decoder was not instructed to parse the
1D/2D indication bit after EOL. This caused later
issues when parsing the succeeding 1D/2D code,
e.g. resulting in the decoder encountering a
negative 1D code and returning with an error.

Setting k>0 means the 1D/2D indication bit is parsed.
The CCITT fax group 3 specification recommends
setting k based on the vertical resolution. MuPDF's
fax decoder depends only on if k is <0, 0 or >0, not
its value (as recommended by the PDF specification).
Therefore it is not necessary to use the vertical
resolution to determine a particular value for k.