Fix key_storable operations leading to leaks of fz_image.
key_storable objects can either be freed 'directly' by an API call
(such as fz_drop_image), or 'internally' by the store realising that
the sole thing holding onto a key_storable is a key.

The current code frees more structure in the direct call than it
does in the internal call. Clearly this is wrong and leads to
leaks.

The fix is to do ALL the freeing in the internal 'drop' function
within the key_storable.

This means we don't need (or want) either fz_drop_key_storable_key or
fz_drop_key_storable to return an int to tell us whether they were
actually dropped, so we make that change to simplify the code.

This shifts the responsibility for freeing the extra internal
structure into the innermost drop functions - fz_drop_image_imp and
fz_drop_image_gprf_imp. To avoid duplicating code, we put the
extra freeing logic into a fz_drop_image_base function.