Fixes for multithreaded
This fixes a few issues.  One issue is that we can not be changing the
input and output formatter for the links if they are being shared among
threads.  This adds the format as part of the link definition.  Next step
is to create a link clone operation in lcms so that we can readily create
a different one with everything the same except the formatter.

Also, this shares the profiles that exist in the color context among the
threads.  When a context is cloned, it will use the profiles in the current
context but it will create a new cmm context.c