Change CMM to use lower resolution tables in links
The table resolution greatly affects the performance for some files.  I don't see any significant color rendering issues going with a lower resolution.

Selecting cmsFLAGS_LOWRESPRECALC uses a MLUT resolution of 17.   This is going to be
sufficient color wise and gives a large improvement performance wise especially for files
where the color link creation is significant compared to the rendering.