Improve PDF color params handling.
PDF allows different OP settings for strokes and fills.

This means that either we need to keep separate entries in
each fz_color_param structure for the stroke one or the fill
one, OR we need to duplicate the fz_color_param structure.

It seems neatest to do the latter (not least because this
means we don't pass more information to each device function
than it actually needs).

Accordingly, we put an fz_color_param in each pdf_material.

We update the code that reads ExtGStates to set the values
appropriately. We take the opportunity to add support for the
PDF 2.0 UseBlackPtComp option too.