If page has no group then use output intent color space
If the document has an output intent but no page group was defined
then we should not let the target device color space define what
our base group is.  Instead we should push a group that uses
the target output intent color space.  This problem occurs on
page 3 of the Altona test file (TextKnockout page).