Fix Gradients in the presence of spots.
Incorporates fixes from Michael: When dealing with spot colors, if the
destination space is pure CMYK and we have unsupported spots we can
do a weighted addition of the spot CMYK colorant like Adobe does
in the PSD image format.   If the destination space is RGB we need
to do a traditional color mapping of each pixel.