Update draw-blend.c to support spot blending.
Proper blending of spots is dependent upon the blending
color space as well as the blend mode.  In particular
when the blend mode is non-separable or non-white preserving
normal blending should be used for the spot colorants.

Incorporates various fixes and optimisations squashed back
to this one commit for clarity. Some of these fixes/optimisations
are due to Michael Vrhel.

In particular we move to handling non-isolated groups in the same
way as gs.