Fix overprint color component detection.
If a colorant is not an explicit separation, then it will be
mapped down to process colorants. These cannot be protected
from overprint.

Incorporates a fix for a problem with shades seen on page 8 of.
Altona_Technical_v20_x4. This shows a problem where we were failing
to correctly overprint a grey gradient with a spot color one.

The overprint setup code was checking for 'Orange' being one of
the process colours, and not finding it. Consequently it was deciding
that Orange would have to be mapped to a process color, and so the
process colors could not be protected from overprint.

The code should have spotted that 'Orange' was one of the spot
colors. With this added, it works correctly.