Clear isolated flag within 'inner' knockout groups.
Knockout groups are only created temporarily, and this solves
problems in Altona_Technical_v20_x4.pdf on page 7 on the 'B'.

It seems reasonable that we shouldn't need to have isolated
enabled here, because it's really required for blending the
whole isolated group back.

While we don't propagate "isolated" into "inner" knockout groups, we
DO need to use the incoming isolated value to correctly establish the
backdrop to use.