Fix knockout operation.
The "blend back" at the end of the inner knockout groups was
attempting to reuse the existing blending code. This was going
wrong for all sorts of reasons (not least the uncomposition
phase) for knockout groups containing alpha, such as found on
page 7 of Altona_Technical_v20_x4.pdf.

Use a dedicated routine. This is much simpler as it doesn't have
to cope with blend modes etc.