gl: Remove GLFW and reinstate FreeGLUT.
GLFW doesn't build on Visual Studio 2005 anymore, and I don't have time to keep
up with the changes.

So, we're switching back to FreeGLUT, which is more stable.

I've added the two missing features that made us switch to GLFW in the first
place: input methods and system clipboard support.

If MuPDF is compiled with our version of FreeGLUT, we now use these functions:

	* glutKeyboardExtFunc
	* glutSetClipboard
	* glutGetClipboard