Fix problem with spot shadings in isolated groups.
As seen with ../tests_private/comparefiles/Bug693541.pdf

This file has an RGB isolated group, within which it renders a
spot only shading. We therefore create the RGB+S+A pixmap, and
set it to all zeros. The shading is drawn to a new S+A pixmap.
The problem comes when the code writes the S+A pixmap to the
RGB+S+A one.

When we set the alpha values to be non zero in an additive space
we need to reset the process components to be full (scaled by
alpha).