Don't cache page count in pdf_document.
Make sure any changes to the page tree are always reflected immediately.

The rev_page_map lookup cache exists when we load the outlines in order
to resolve links faster, so we don't need to worry about that one.

The linear_page_refs stuff is more troublesome, so don't mix editing a
PDF with progressive loading!