Fix multithreaded crash with tiled regions.
Michael has found a crash when scrolling quickly through pages
with gsview. 2 Threads are redrawing at the same time from a
display list. The problem comes when both threads happen to be
trying to draw the same tile from the cache at the same time.

The current code alters the ->{x,y} values of the pixmap from
the cache as it tiles. If 2 threads are using the same tile
at the same time, this causes a race condition which can upset
the clipping calculations and we can access out of range.

The solution is to make a new 'wrapper' fz_pixmap around the
same data, and to alter the x/y values there instead.

We therefore introduce a (hopefully generally useful) function
fz_new_pixmap_from_pixmap, and use that.