Fix pgm corruption with spots files.
When converting from a source space to a destination space with
spots for our "overprint" group push, we were hitting problems
in our use of the pixmap converters.

Pixmap converters with copy_spots set can assume that the source
and destination spots are the same - when copy spots is NOT set,
we cannot assume this. This was leading us to have uninitialised
group backgrounds.

I believe we were only seeing this with pgm because of the device
k to cmyk as K only special case.

Also fix an error in the fast_gray_to_cmyk routine that failed to
account for the change from subtractive to additive.