Bug 689699: Avoid buffer overrun.
When cleaning a pdf file, various lists (of pdf_xref_len length) are
defined early on.

If we trigger a repair during the clean, this can cause pdf_xref_len
to increase causing an overrun.

Fix this by watching for changes in the length, and checking accesses
to the list for validity.

This also appears to fix bugs 698700-698703.