Avoid revisiting the same page tree node repeatedly.
Only attempt to progress to new page tree nodes if a new node to
process has actually been found. Previously if e.g. doc.findPage(1)
was called in murun for a single page document the error message was
"cycle in page tree" because the single page tree node was visited
twice. Now the error message is "cannot find page 1 in page tree",
which is more accurate.