Support CalRGB (and CalGray) as destination color spaces
I did not foresee the case where a transparency groups color
space could be a Cal color space.  I had always thought of them
as only be source color spaces not destination color spaces.
This commit makes sure that we have the equivalent ICC profile
when the destination is a Cal space.