Fix 699332: Copy resources when cleaning content streams.
Previously if both cleaning and sanitizing content streams the
pages' resource dictionaries would retain the actually used
resources. If the content streams were only cleaned and not
sanitized the page's resource dictionaries were incorrectly
emptied. All resources, whether used or not, ought to be
retained, as is the case after this commit.