Use bin2coff to generate font data blobs object files.
This is the windows equivalent of objcopy or ld -r binary.

We need different bin2coff font object files for 32 and 64 bit builds,
so put the font object files in AdditionalDependencies.

Revamp Windows builds to use bin2coff.

Fix bin2coff to not use '-' in symbol names.

Add a new project file to make bin2coff.

Add a .rules file to tell MSVC how to call the built version of bin2coff.

Update libresources build to use this rules file.

Update noto.c to correctly find the symbols we make.