MSVC solution fixes.
Having libluratech set to not build in non-commercial builds is
not enough to make stuff work, as the dependencies will still
try to link it in.

Instead, we now have a Release configuration for libluratech
with every file disabled, and the non-commercial builds depend
on that.

Also, tweak bin2coff build to put it's buildlog in the usual
place.