Fix MSVC build issues.
When updating from VS2005 to VS2017, the .rules mechanism is
converted to a .props based scheme. This is not entirely
equivalent, and so causes problems with our "build both
32 and 64bit resources" scheme.

Instead, we amend the flag handling in bin2coff so that it
will accept "Win32" and "x64" as flags. These can then be driven
easily from the project files, giving us a "more standard"
way of working.

Also, remove some stray fonts that had crept into the resources
build.