Rejig pdf_update_page and pdf_update_annot.
The intent is for a user to iterate over the annotations on a page calling
pdf_update_annot for each one. If this function returns true, then the
annotation has changed since the last time it was called, and the user
needs to re-render.

pdf_update_page is a simple loop over the annotations on a page, for use
if you only care about page level granularity.

Users should no longer look at or change the pdf_annot.has_new_ap field.